requirements.txt
Django
djangorestframework
django-filter
drf-yasg
celery
django-oauth-toolkit
social-auth-app-django
django-rest-auth
django-cors-headers
python-dotenv
django-celery-results
redis
psycopg2-binary
openai
gunicorn
django-storages
boto3
pycurl
pillow
langchain
langchain-community
langchain-openai
pinecone
bs4
tiktoken
pypdf
requests
beautifulsoup4
unstructured